requirements.txt
pandas>=2.2
pyyaml>=6.0
lightgbm>=4.0
qlib>=0.9.7
scikit-learn>=1.5
requests>=2.31
pypdf>=5.0
akshare>=1.18
